[04/09/2012] modified   examples/ApplicationIT_regularvector/src/MainIT_regvector.cpp 1. [ADD] Do not declare an array of variable size for MSVC.
modified   src/ITL_histogram.cpp
1. [ADD] Correct a statement in the function for cross validation.



git-svn-id: https://svn.mcs.anl.gov/repos/itl/trunk@67 609b6239-7616-4a8b-b317-616138058241

+++ src/ITL_histogram.cpp
@@ -320,8 +320,7 @@ ITL_histogram::crossValidateSpeedUp( ITL_field_regular<SCALAR> *scalarField, cha
 
 	for( int i = 0; i<nIter; i++ )
 	{
-		nBinArray[nIter-i-1] = (int)(nMax/(pow(2.0,(float)i)));
-
+		nBinArray[nIter-i-1] = (int)((float)nMax/powf(2.0f,(float)i));	// MOD-BY-LEETEN 04/09/2012-FROM:		nBinArray[nIter-i-1] = (int)(nMax/(pow(2.0,(float)i)));
 	}
 
 	for(int i=0;i<nIter;i++)
